ui-components: Input component
- TextInput component
- NumberInput component
- BaseInput - can be used to compose own
type of Input component.
