--- packages/ui-components/src/Input/Input.module.scss
@@ -1,89 +1,0 @@
-@import "../styles/tokens.scss";
-
-$input-side-padding: crl-gutters(1.5);
-$input-min-width: 160px - $input-side-padding;
-
-.input {
-  height: 40px;
-  border-radius: 3px;
-  font-family: Source Sans Pro, sans-serif; // TODO (koorosh): replace with token value
-  font-size: 14px; // TODO (koorosh): replace with token value (default font size)
-  line-height: 22px;
-  border: solid 1px $crl-neutral-4;
-  background-color: $crl-base-white;
-  padding: 0 $input-side-padding;
-  position: relative;
-  min-width: $input-min-width;
-
-  &:focus {
-    border: solid 1px $crl-base-blue;
-  }
-
-  &:hover {
-    border: solid 1px $crl-neutral-5;
-  }
-}
-
-.invalid {
-  border: solid 1px $crl-base-red;
-  background-color: $crl-red-1;
-  color: $crl-base-red;
-
-  &:focus, &:hover {
-    border: solid 1px $crl-base-red;
-  }
-}
-
-.disabled {
-  background: $crl-neutral-1;
-  color: $crl-base-text--light;
-  border: solid 1px $crl-neutral-3;
-  cursor: default;
-
-  &:focus, &:hover {
-    border: solid 1px $crl-neutral-3;
-  }
-}
-
-// remove default arrow spinner for input (number type)
-.input[type="number"]::-webkit-outer-spin-button,
-.input[type="number"]::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-  pointer-events: none;
-}
-
-// remove default arrow spinner for input -- Firefox browser
-.input[type=number] {
-  -moz-appearance: textfield;
-}
-
-// NumberInput styles
-.input-number-wrapper {
-  display: flex;
-  flex-direction: row;
-
-  &> input {
-    padding-right: 24px;
-  }
-}
-
-.spin-buttons-group {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  position: relative;
-  right: 20px;
-  width: 20px;
-  margin: 8px 0;
-}
-
-.spin-button {
-  display: flex;
-  flex-grow: 1;
-  align-items: center;
-
-  &> svg {
-    width: 8px;
-  }
-}
