+++ packages/ui-components/src/Input/index.ts
@@ -1,1 +1,3 @@
-export * from "./Input";
+export * from "./BaseInput";
+export * from "./TextInput";
+export * from "./NumberInput";
