+++ packages/ui-components/src/Input/styles.module.scss
@@ -1,0 +1,2 @@
+@import "./input.scss";
+@import "./number-input.scss";
